move gp into gps
