+++ bofire/models/gps/__init__.py
@@ -1,0 +1,3 @@
+from bofire.models.gps.gps import MixedSingleTaskGPModel, SingleTaskGPModel
+
+__all__ = ["MixedSingleTaskGPModel", "SingleTaskGPModel"]
