+++ bofire/strategies/botorch/base.py
@@ -26,11 +26,8 @@ from bofire.domain.features import (
     OutputFeatures,
     TInputTransformSpecs,
 )
-from bofire.models.torch_models import (
-    BotorchModels,
-    MixedSingleTaskGPModel,
-    SingleTaskGPModel,
-)
+from bofire.models.gps import MixedSingleTaskGPModel, SingleTaskGPModel
+from bofire.models.torch_models import BotorchModels
 from bofire.strategies.strategy import PredictiveStrategy
 from bofire.strategies.utils import is_power_of_two
 from bofire.utils.enum import (  # DescriptorMethodEnum,
