+++ tests/bofire/models/gps/test_gps.py
@@ -1,0 +1,105 @@
+import pytest
+import torch
+from botorch.models import MixedSingleTaskGP, SingleTaskGP
+from botorch.models.transforms.input import (
+    ChainedInputTransform,
+    InputStandardize,
+    Normalize,
+)
+from botorch.models.transforms.outcome import Standardize
+
+from bofire.domain.features import (
+    CategoricalInput,
+    ContinuousInput,
+    ContinuousOutput,
+    InputFeatures,
+    OutputFeatures,
+)
+from bofire.models.gps import MixedSingleTaskGPModel, SingleTaskGPModel
+from bofire.models.gps.kernels import HammondDistanceKernel, RBFKernel, ScaleKernel
+from bofire.utils.enum import CategoricalEncodingEnum, ScalerEnum
+from bofire.utils.torch_tools import OneHotToNumeric
+
+
+@pytest.mark.parametrize(
+    "kernel, scaler",
+    [
+        (ScaleKernel(base_kernel=RBFKernel(ard=True)), ScalerEnum.NORMALIZE),
+        (ScaleKernel(base_kernel=RBFKernel(ard=False)), ScalerEnum.STANDARDIZE),
+    ],
+)
+def test_SingleTaskGPModel(kernel, scaler):
+    input_features = InputFeatures(
+        features=[
+            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
+            for i in range(2)
+        ]
+    )
+    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
+    experiments = input_features.sample(n=10)
+    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
+    experiments["valid_y"] = 1
+    model = SingleTaskGPModel(
+        input_features=input_features,
+        output_features=output_features,
+        kernel=kernel,
+        scaler=scaler,
+    )
+    model.fit(experiments)
+    preds = model.predict(experiments)
+    assert preds.shape == (10, 2)
+    # check that model is composed correctly
+    assert isinstance(model.model, SingleTaskGP)
+    assert isinstance(model.model.outcome_transform, Standardize)
+    if scaler == ScalerEnum.NORMALIZE:
+        assert isinstance(model.model.input_transform, Normalize)
+    else:
+        assert isinstance(model.model.input_transform, InputStandardize)
+
+
+@pytest.mark.parametrize(
+    "kernel, scaler",
+    [
+        (RBFKernel(ard=True), ScalerEnum.NORMALIZE),
+        (RBFKernel(ard=False), ScalerEnum.STANDARDIZE),
+    ],
+)
+def test_MixedGPModel(kernel, scaler):
+    input_features = InputFeatures(
+        features=[
+            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
+            for i in range(2)
+        ]
+        + [CategoricalInput(key="x_cat", categories=["mama", "papa"])]
+    )
+    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
+    experiments = input_features.sample(n=10)
+    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
+    experiments.loc[experiments.x_cat == "mama", "y"] *= 5.0
+    experiments.loc[experiments.x_cat == "papa", "y"] /= 2.0
+    experiments["valid_y"] = 1
+
+    model = MixedSingleTaskGPModel(
+        input_features=input_features,
+        output_features=output_features,
+        input_preprocessing_specs={"x_cat": CategoricalEncodingEnum.ONE_HOT},
+        scaler=scaler,
+        continuous_kernel=kernel,
+        categorical_kernel=HammondDistanceKernel(),
+    )
+
+    model.fit(experiments)
+    preds = model.predict(experiments)
+    assert preds.shape == (10, 2)
+    # check that model is composed correctly
+    assert isinstance(model.model, MixedSingleTaskGP)
+    assert isinstance(model.model.outcome_transform, Standardize)
+    assert isinstance(model.model.input_transform, ChainedInputTransform)
+    if scaler == ScalerEnum.NORMALIZE:
+        assert isinstance(model.model.input_transform.tf1, Normalize)
+    else:
+        assert isinstance(model.model.input_transform.tf1, InputStandardize)
+    assert torch.eq(
+        model.model.input_transform.tf1.indices, torch.tensor([0, 1], dtype=torch.int64)
+    ).all()
+    assert isinstance(model.model.input_transform.tf2, OneHotToNumeric)
