+++ tests/bofire/models/test_feature_importance.py
@@ -13,7 +13,7 @@ from bofire.models.feature_importance import (
     permutation_importance,
     permutation_importance_hook,
 )
-from bofire.models.torch_models import SingleTaskGPModel
+from bofire.models.gps import SingleTaskGPModel
 
 
 def get_model_and_data():
