+++ tests/bofire/models/test_torch_models.py
@@ -1,4 +1,3 @@
-import gpytorch.priors
 import numpy as np
 import pandas as pd
 import pytest
@@ -11,8 +10,6 @@ from botorch.models.transforms.input import (
     InputStandardize,
     Normalize,
 )
-from botorch.models.transforms.outcome import Standardize
-from gpytorch.kernels import LinearKernel, MaternKernel, RBFKernel
 from torch import Tensor
 
 from bofire.domain.features import (
@@ -23,76 +20,10 @@ from bofire.domain.features import (
     InputFeatures,
     OutputFeatures,
 )
-from bofire.models.priors import GammaPrior
-from bofire.models.torch_models import (
-    RBF,
-    BotorchModels,
-    ContinuousKernel,
-    EmpiricalModel,
-    HammondDistanceKernel,
-    Linear,
-    Matern,
-    MixedSingleTaskGPModel,
-    SingleTaskGPModel,
-)
+from bofire.models.gps import MixedSingleTaskGPModel, SingleTaskGPModel
+from bofire.models.torch_models import BotorchModels, EmpiricalModel
 from bofire.utils.enum import CategoricalEncodingEnum, ScalerEnum
 from bofire.utils.torch_tools import OneHotToNumeric, tkwargs
-
-
-@pytest.mark.parametrize(
-    "kernel, ard_num_dims, active_dims, expected_kernel",
-    [
-        (
-            RBF(ard=False, lengthscale_prior=GammaPrior(concentration=2.0, rate=0.15)),
-            10,
-            list(range(5)),
-            RBFKernel,
-        ),
-        (
-            RBF(ard=False),
-            10,
-            list(range(5)),
-            RBFKernel,
-        ),
-        (RBF(ard=True), 10, list(range(5)), RBFKernel),
-        (
-            Matern(
-                ard=False, lengthscale_prior=GammaPrior(concentration=2.0, rate=0.15)
-            ),
-            10,
-            list(range(5)),
-            MaternKernel,
-        ),
-        (Matern(ard=False), 10, list(range(5)), MaternKernel),
-        (Matern(ard=True), 10, list(range(5)), MaternKernel),
-        (Matern(ard=False, nu=2.5), 10, list(range(5)), MaternKernel),
-        (Matern(ard=True, nu=1.5), 10, list(range(5)), MaternKernel),
-        (Linear(), 10, list(range(5)), LinearKernel),
-    ],
-)
-def test_continuous_kernel(
-    kernel: ContinuousKernel, ard_num_dims, active_dims, expected_kernel
-):
-    k = kernel.to_gpytorch(
-        batch_shape=torch.Size(), ard_num_dims=ard_num_dims, active_dims=active_dims
-    )
-    assert isinstance(k, expected_kernel)
-    if isinstance(kernel, Linear):
-        return
-    if kernel.lengthscale_prior is not None:
-        assert hasattr(k, "lengthscale_prior")
-        assert isinstance(k.lengthscale_prior, gpytorch.priors.GammaPrior)
-    else:
-        assert hasattr(k, "lengthscale_prior") is False
-
-    if kernel.ard is False:
-        assert k.ard_num_dims is None
-    else:
-        assert k.ard_num_dims == len(active_dims)
-    assert torch.eq(k.active_dims, torch.tensor(active_dims, dtype=torch.int64)).all()
-
-    if isinstance(kernel, Matern):
-        assert kernel.nu == k.nu
 
 
 @pytest.mark.parametrize("modelclass", [(SingleTaskGPModel), (MixedSingleTaskGPModel)])
@@ -179,249 +110,6 @@ def test_BotorchModel_validate_invalid_input_preprocessing_steps(
         )
 
 
-@pytest.mark.parametrize(
-    "kernel, scaler",
-    [(RBF(ard=True), ScalerEnum.NORMALIZE), (RBF(ard=False), ScalerEnum.STANDARDIZE)],
-)
-def test_SingleTaskGPModel(kernel, scaler):
-    input_features = InputFeatures(
-        features=[
-            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
-            for i in range(2)
-        ]
-    )
-    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
-    experiments = input_features.sample(n=10)
-    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
-    experiments["valid_y"] = 1
-    model = SingleTaskGPModel(
-        input_features=input_features,
-        output_features=output_features,
-        kernel=kernel,
-        scaler=scaler,
-    )
-    model.fit(experiments)
-    preds = model.predict(experiments)
-    assert preds.shape == (10, 2)
-    # check that model is composed correctly
-    assert isinstance(model.model, SingleTaskGP)
-    assert isinstance(model.model.outcome_transform, Standardize)
-    if scaler == ScalerEnum.NORMALIZE:
-        assert isinstance(model.model.input_transform, Normalize)
-    else:
-        assert isinstance(model.model.input_transform, InputStandardize)
-
-
-@pytest.mark.parametrize("folds", [5, 3, 10, -1])
-def test_model_cross_validate(folds):
-    input_features = InputFeatures(
-        features=[
-            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
-            for i in range(2)
-        ]
-    )
-    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
-    experiments = input_features.sample(n=100)
-    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
-    experiments["valid_y"] = 1
-    experiments = experiments.sample(10)
-    model = SingleTaskGPModel(
-        input_features=input_features,
-        output_features=output_features,
-    )
-    train_cv, test_cv, _ = model.cross_validate(experiments, folds=folds)
-    efolds = folds if folds != -1 else 10
-    assert len(train_cv.results) == efolds
-    assert len(test_cv.results) == efolds
-
-
-def test_model_cross_validate_descriptor():
-    folds = 5
-    input_features = InputFeatures(
-        features=[
-            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
-            for i in range(2)
-        ]
-        + [
-            CategoricalDescriptorInput(
-                key="x_3",
-                categories=["a", "b", "c"],
-                descriptors=["alpha"],
-                values=[[1], [2], [3]],
-            )
-        ]
-    )
-    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
-    experiments = input_features.sample(n=100)
-    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
-    experiments.loc[experiments.x_2 == "b", "y"] += 5
-    experiments.loc[experiments.x_2 == "c", "y"] += 10
-    experiments["valid_y"] = 1
-    experiments = experiments.sample(10)
-    for encoding in [
-        CategoricalEncodingEnum.ONE_HOT,
-        CategoricalEncodingEnum.DESCRIPTOR,
-    ]:
-        model = SingleTaskGPModel(
-            input_features=input_features,
-            output_features=output_features,
-            input_preprocessing_specs={"x_3": encoding},
-        )
-        train_cv, test_cv, _ = model.cross_validate(experiments, folds=folds)
-        efolds = folds if folds != -1 else 10
-        assert len(train_cv.results) == efolds
-        assert len(test_cv.results) == efolds
-
-
-@pytest.mark.parametrize("include_X", [True, False])
-def test_model_cross_validate_include_X(include_X):
-    input_features = InputFeatures(
-        features=[
-            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
-            for i in range(2)
-        ]
-    )
-    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
-    experiments = input_features.sample(n=10)
-    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
-    experiments["valid_y"] = 1
-    model = SingleTaskGPModel(
-        input_features=input_features,
-        output_features=output_features,
-    )
-    train_cv, test_cv, _ = model.cross_validate(
-        experiments, folds=5, include_X=include_X
-    )
-    if include_X:
-        assert train_cv.results[0].X.shape == (8, 2)
-        assert test_cv.results[0].X.shape == (2, 2)
-    if include_X is False:
-        assert train_cv.results[0].X is None
-        assert test_cv.results[0].X is None
-
-
-def test_model_cross_validate_hooks():
-    def hook1(model, X_train, y_train, X_test, y_test):
-        assert isinstance(model, SingleTaskGPModel)
-        assert y_train.shape == (8, 1)
-        assert y_test.shape == (2, 1)
-        return X_train.shape
-
-    def hook2(model, X_train, y_train, X_test, y_test, return_test=True):
-        if return_test:
-            return X_test.shape
-        return X_train.shape
-
-    input_features = InputFeatures(
-        features=[
-            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
-            for i in range(2)
-        ]
-    )
-    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
-    experiments = input_features.sample(n=10)
-    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
-    experiments["valid_y"] = 1
-    #
-    model = SingleTaskGPModel(
-        input_features=input_features,
-        output_features=output_features,
-    )
-    # first test with one hook
-    _, _, hook_results = model.cross_validate(
-        experiments, folds=5, hooks={"hook1": hook1}
-    )
-    assert len(hook_results.keys()) == 1
-    assert len(hook_results["hook1"]) == 5
-    assert hook_results["hook1"] == [(8, 2), (8, 2), (8, 2), (8, 2), (8, 2)]
-    # now test with two hooks
-    _, _, hook_results = model.cross_validate(
-        experiments, folds=5, hooks={"hook1": hook1, "hook2": hook2}
-    )
-    assert len(hook_results.keys()) == 2
-    assert len(hook_results["hook1"]) == 5
-    assert hook_results["hook1"] == [(8, 2), (8, 2), (8, 2), (8, 2), (8, 2)]
-    assert len(hook_results["hook2"]) == 5
-    assert hook_results["hook2"] == [(2, 2), (2, 2), (2, 2), (2, 2), (2, 2)]
-    # now test with two hooks and keyword arguments
-    _, _, hook_results = model.cross_validate(
-        experiments,
-        folds=5,
-        hooks={"hook1": hook1, "hook2": hook2},
-        hook_kwargs={"hook2": {"return_test": False}},
-    )
-    assert len(hook_results.keys()) == 2
-    assert len(hook_results["hook1"]) == 5
-    assert hook_results["hook1"] == [(8, 2), (8, 2), (8, 2), (8, 2), (8, 2)]
-    assert len(hook_results["hook2"]) == 5
-    assert hook_results["hook2"] == [(8, 2), (8, 2), (8, 2), (8, 2), (8, 2)]
-
-
-@pytest.mark.parametrize("folds", [-2, 0, 1, 11])
-def test_model_cross_validate_invalid(folds):
-    input_features = InputFeatures(
-        features=[
-            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
-            for i in range(2)
-        ]
-    )
-    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
-    experiments = input_features.sample(n=10)
-    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
-    experiments["valid_y"] = 1
-    model = SingleTaskGPModel(
-        input_features=input_features,
-        output_features=output_features,
-    )
-    with pytest.raises(ValueError):
-        model.cross_validate(experiments, folds=folds)
-
-
-@pytest.mark.parametrize(
-    "kernel, scaler",
-    [(RBF(ard=True), ScalerEnum.NORMALIZE), (RBF(ard=False), ScalerEnum.STANDARDIZE)],
-)
-def test_MixedGPModel(kernel, scaler):
-    input_features = InputFeatures(
-        features=[
-            ContinuousInput(key=f"x_{i+1}", lower_bound=-4, upper_bound=4)
-            for i in range(2)
-        ]
-        + [CategoricalInput(key="x_cat", categories=["mama", "papa"])]
-    )
-    output_features = OutputFeatures(features=[ContinuousOutput(key="y")])
-    experiments = input_features.sample(n=10)
-    experiments.eval("y=((x_1**2 + x_2 - 11)**2+(x_1 + x_2**2 -7)**2)", inplace=True)
-    experiments.loc[experiments.x_cat == "mama", "y"] *= 5.0
-    experiments.loc[experiments.x_cat == "papa", "y"] /= 2.0
-    experiments["valid_y"] = 1
-
-    model = MixedSingleTaskGPModel(
-        input_features=input_features,
-        output_features=output_features,
-        input_preprocessing_specs={"x_cat": CategoricalEncodingEnum.ONE_HOT},
-        scaler=scaler,
-        continuous_kernel=kernel,
-        categorical_kernel=HammondDistanceKernel(),
-    )
-
-    model.fit(experiments)
-    preds = model.predict(experiments)
-    assert preds.shape == (10, 2)
-    # check that model is composed correctly
-    assert isinstance(model.model, MixedSingleTaskGP)
-    assert isinstance(model.model.outcome_transform, Standardize)
-    assert isinstance(model.model.input_transform, ChainedInputTransform)
-    if scaler == ScalerEnum.NORMALIZE:
-        assert isinstance(model.model.input_transform.tf1, Normalize)
-    else:
-        assert isinstance(model.model.input_transform.tf1, InputStandardize)
-    assert torch.eq(
-        model.model.input_transform.tf1.indices, torch.tensor([0, 1], dtype=torch.int64)
-    ).all()
-    assert isinstance(model.model.input_transform.tf2, OneHotToNumeric)
-
-
 def test_BotorchModels_invalid_output_features():
     model1 = SingleTaskGPModel(
         input_features=InputFeatures(
@@ -431,7 +119,6 @@ def test_BotorchModels_invalid_output_features():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
     )
     model2 = SingleTaskGPModel(
@@ -442,7 +129,6 @@ def test_BotorchModels_invalid_output_features():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
     )
     with pytest.raises(ValueError):
@@ -458,7 +144,6 @@ def test_BotorchModels_invalid_input_features():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
     )
     model2 = SingleTaskGPModel(
@@ -470,7 +155,6 @@ def test_BotorchModels_invalid_input_features():
             + [CategoricalInput(key="x_3", categories=["apple", "banana"])]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
     )
     with pytest.raises(ValueError):
@@ -494,7 +178,6 @@ def test_BotorchModels_invalid_preprocessing():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
         input_preprocessing_specs={"cat": CategoricalEncodingEnum.ONE_HOT},
     )
@@ -514,7 +197,6 @@ def test_BotorchModels_invalid_preprocessing():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y2")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
         input_preprocessing_specs={"cat": CategoricalEncodingEnum.DESCRIPTOR},
     )
@@ -547,7 +229,6 @@ def test_BotorchModels_invalid_preprocessing():
                     output_features=OutputFeatures(
                         features=[ContinuousOutput(key="y")]
                     ),
-                    kernel=RBF(),
                     scaler=ScalerEnum.NORMALIZE,
                     input_preprocessing_specs={"cat": CategoricalEncodingEnum.ONE_HOT},
                 ),
@@ -574,7 +255,6 @@ def test_BotorchModels_invalid_preprocessing():
                             ContinuousOutput(key="y3"),
                         ]
                     ),
-                    kernel=RBF(),
                     scaler=ScalerEnum.NORMALIZE,
                     input_preprocessing_specs={"cat": CategoricalEncodingEnum.ONE_HOT},
                 ),
@@ -612,7 +292,6 @@ def test_botorch_models_invalid_number_of_outputs(models):
                     output_features=OutputFeatures(
                         features=[ContinuousOutput(key="y")]
                     ),
-                    kernel=RBF(),
                     scaler=ScalerEnum.NORMALIZE,
                     input_preprocessing_specs={"cat": CategoricalEncodingEnum.ONE_HOT},
                 ),
@@ -636,7 +315,6 @@ def test_botorch_models_invalid_number_of_outputs(models):
                     output_features=OutputFeatures(
                         features=[ContinuousOutput(key="y2")]
                     ),
-                    kernel=RBF(),
                     scaler=ScalerEnum.NORMALIZE,
                     input_preprocessing_specs={"cat": CategoricalEncodingEnum.ONE_HOT},
                 ),
@@ -665,7 +343,6 @@ def test_botorch_models_check_compatibility():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
         input_preprocessing_specs={"cat": CategoricalEncodingEnum.ONE_HOT},
     )
@@ -685,7 +362,6 @@ def test_botorch_models_check_compatibility():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y2")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
         input_preprocessing_specs={"cat": CategoricalEncodingEnum.ONE_HOT},
     )
@@ -810,7 +486,6 @@ def test_botorch_models_input_preprocessing_specs():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
         input_preprocessing_specs={"cat": CategoricalEncodingEnum.DESCRIPTOR},
     )
@@ -828,7 +503,6 @@ def test_botorch_models_input_preprocessing_specs():
             ]
         ),
         output_features=OutputFeatures(features=[ContinuousOutput(key="y2")]),
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
         input_preprocessing_specs={"cat2": CategoricalEncodingEnum.ONE_HOT},
     )
@@ -854,7 +528,6 @@ def test_botorch_models_fit_and_compatibilize():
     model1 = SingleTaskGPModel(
         input_features=input_features,
         output_features=output_features,
-        kernel=RBF(),
         scaler=ScalerEnum.NORMALIZE,
     )
     # model 2
@@ -878,8 +551,6 @@ def test_botorch_models_fit_and_compatibilize():
         output_features=output_features,
         input_preprocessing_specs={"x_cat": CategoricalEncodingEnum.ONE_HOT},
         scaler=ScalerEnum.STANDARDIZE,
-        continuous_kernel=Matern(nu=2.5),
-        categorical_kernel=HammondDistanceKernel(),
     )
     # create models
     models = BotorchModels(models=[model1, model2])
@@ -985,8 +656,6 @@ def test_empirical_model():
         output_features=output_features,
         input_preprocessing_specs={"x_cat": CategoricalEncodingEnum.ONE_HOT},
         scaler=ScalerEnum.STANDARDIZE,
-        continuous_kernel=Matern(nu=2.5),
-        categorical_kernel=HammondDistanceKernel(),
     )
     # create models
     models = BotorchModels(models=[model1, model2])
